moved items from lib → internals; better look for docs
