+++ src/tools/interpreter.rs
@@ -1,5 +1,5 @@
 use colored::Colorize;
-use stck::{api, error::Error};
+use stck::prelude::*;
 
 #[derive(PartialEq, Clone, Copy)]
 enum StckMode {
@@ -10,14 +10,14 @@ enum StckMode {
     PrintProccCode,
 }
 
-fn print_code(code: &stck::Code, import_stack: usize) {
+fn print_code(code: &stck::internals::Code, import_stack: usize) {
     for expr in code {
         if import_stack != 0 {
             println!("{} {}", ">".repeat(import_stack).blue(), expr.cont);
         } else {
             println!("{}", expr.cont);
         }
-        if let stck::ExprCont::IncludedCode(code) = &expr.cont {
+        if let stck::internals::ExprCont::IncludedCode(code) = &expr.cont {
             print_code(code, import_stack + 1);
         }
     }
@@ -27,20 +27,20 @@ fn execute(mode: StckMode, file_path: String) -> Result<(), Error> {
     use StckMode as M;
     match mode {
         M::Normal => {
-            api::execute_file(file_path)?;
+            execute_file(file_path)?;
         }
         M::Debug => {
             todo!()
         }
         M::SyntaxCheck => {
-            api::get_project_code(file_path)?;
+            get_project_code(file_path)?;
         }
         M::PrintProccCode => {
-            let code = api::get_project_code(file_path)?;
+            let code = get_project_code(file_path)?;
             print_code(&code, 0);
         }
         M::TokenCheck => {
-            api::get_tokens(file_path)?;
+            get_tokens(file_path)?;
         }
     }
     Ok(())
